requirements.txt
fastapi==0.89.1
pandas==1.5.3
pandas_stubs==1.2.0.35
pdf2image==1.16.2
Pillow==9.4.0
pydantic==1.8.2
python_Levenshtein==0.20.9
regex==2022.10.31
streamlit==1.17.0
uvicorn==0.20.0
